Move print_last_interpreted_history().
